chore: relocate style components into src
